moved directory and class

+++ src/main/java/com/devho/javatoy/repository/UserProjectRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProjectRepository extends JpaRepository<UserProject, Long>{
-
+public interface UserProjectRepository extends JpaRepository<UserProject, Integer>{
 }
